Improvement - changes in params structure in matchType operator
