+++ packages/core/src/operators/index.ts
@@ -1,0 +1,2 @@
+export { matchPath } from './matchPath/matchPath.operator';
+export { use } from './use/use.operator';
