+++ packages/core/src/operators/matchPath/matchType.operator.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs';
 import { filter } from 'rxjs/operators';
-import { HttpMethod, HttpRequest } from '../http.interface';
+import { HttpMethod, HttpRequest } from '../../http.interface';
 
 export const matchType = (method: HttpMethod) => (source$: Observable<HttpRequest>) =>
   source$.pipe(
