+++ packages/core/src/operators/use/use.operator.ts
@@ -1,7 +1,7 @@
 import { Observable, of } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
-import { Effect } from '../effects/effects.interface';
-import { HttpRequest, HttpResponse } from '../http.interface';
+import { Effect } from '../../effects/effects.interface';
+import { HttpRequest, HttpResponse } from '../../http.interface';
 
 export const use =
   (middleware: Effect<HttpRequest>, res?: HttpResponse) =>
